requirements.txt
DrissionPage
pydub
SpeechRecognition
Flask
Flask-CORS
requests
python-dotenv